requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
filelock==3.0.12
idna==2.10
py3-validate-email==0.2.13
pycodestyle==2.6.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
